// extensions/ai-commands/src/lib/settings/enum.ts
export enum CommandAnswer {
  ASK_SELECTED_TEXT = "ask-selected-text",
  ASK_WEBPAGE = "ask-webpage",
  CASUAL = "casual",
  CODE_EXPLAIN = "codeexplain",
  CONFIDENT = "confident",
  EXPLAIN = "explain",
  FIX = "fix",
  FRIENDLY = "friendly",
  IMAGE_DESCRIBE = "imagedescribe",
  IMAGE_TO_TEXT = "image-to-text",
  IMPROVE = "improve",
  LONGER = "longer",
  PROFESSIONAL = "professional",
  SHORTER = "shorter",
  TRANSLATE = "translate",
  TWEET = "tweet",
  BROWSER_SUMMARIZE = "browser-summarize",
}